Update scripts.js
created a new variable (pokemonRepository) to hold an IIFE that wraps around pokemonList array to prevent accidently accessing it's global state. The IFFE will now return an object with two keys, "add" (used to add an item to the array) and "getAll"(returns pokemonList array). Updated the forEach loop to cope with these changes.