Update scripts.js
deleted document.write contents and cleared out the forEach loop block. Added variables and used queryselector to select the <ul> element, created a <li> and <button> element using document.createElement function. Added innerText , and a class to the button as well. Lastly, appended the button to the <li> tag as it's child, and the <li> to the <ul> tag as its child.